stream: make loadStream reloadable on any mediaSource
